fix chatglm3 issue
